mipmap

+++ Include/Sampler/SequenceGenerator.hpp
@@ -12,7 +12,7 @@ CUDAInline float radicalInverse(unsigned int bits) {
     bits = ((bits & 0x33333333u) << 2u) | ((bits & 0xCCCCCCCCu) >> 2u);
     bits = ((bits & 0x0F0F0F0Fu) << 4u) | ((bits & 0xF0F0F0F0u) >> 4u);
     bits = ((bits & 0x00FF00FFu) << 8u) | ((bits & 0xFF00FF00u) >> 8u);
-    return static_cast<float>(bits) * 2.3283064365386963e-10f; //0x100000000
+    return bits * 2.3283064365386963e-10f; //0x100000000
 }
 CUDAInline vec2 hammersley(unsigned int i, unsigned int N) {
     return vec2(static_cast<float>(i) / static_cast<float>(N), radicalInverse(i));
